Add full-screen button to the preview modal
