+++ src/editor/library/container.js
@@ -12,10 +12,10 @@ import { CheckboxControl } from '@wordpress/components';
 import { Button, ToggleSwitch } from '../../components';
 import { updateSCOnEditor } from '../../extensions/style-cards';
 import {
-	imageUploader,
 	svgAttributesReplacer,
 	svgCurrentColorStatus,
 	fitSvg,
+	onRequestInsertPattern,
 } from './util';
 import { injectImgSVG, generateDataObject } from '../../extensions/svg';
 import MaxiModal from './modal';
@@ -81,6 +81,7 @@ const MasonryItem = props => {
 							type='preview'
 							url={demoUrl}
 							title={serial}
+							onRequestInsert={onRequestInsert}
 						/>
 					)}
 					<Button
@@ -321,109 +322,109 @@ const LibraryContainer = props => {
 	const [isChecked, setChecked] = useState(false);
 
 	/** Patterns / Blocks */
-	const onRequestInsertPattern = (parsedContent, usePlaceholderImage) => {
-		const isValid = isValidTemplate(parsedContent);
+	// const onRequestInsertPattern = (parsedContent, usePlaceholderImage) => {
+	// 	const isValid = isValidTemplate(parsedContent);
 
-		if (isValid) {
-			const loadingMessage = `<h3>${__(
-				'LOADING…',
-				'maxi-blocks'
-			)}<span class="maxi-spinner"></span></h3>`;
+	// 	if (isValid) {
+	// 		const loadingMessage = `<h3>${__(
+	// 			'LOADING…',
+	// 			'maxi-blocks'
+	// 		)}<span class="maxi-spinner"></span></h3>`;
 
-			onSelect({ content: loadingMessage });
+	// 		onSelect({ content: loadingMessage });
 
-			onRequestClose();
+	// 		onRequestClose();
 
-			const imagesLinks = [];
-			const imagesIds = [];
+	// 		const imagesLinks = [];
+	// 		const imagesIds = [];
 
-			const allImagesRegexp = new RegExp('mediaID":(.*)",', 'g');
+	// 		const allImagesRegexp = new RegExp('mediaID":(.*)",', 'g');
 
-			const allImagesLinks = parsedContent.match(allImagesRegexp);
+	// 		const allImagesLinks = parsedContent.match(allImagesRegexp);
 
-			const allImagesLinksParsed = allImagesLinks.map(image => {
-				const parsed = image.replace(/\\/g, '');
+	// 		const allImagesLinksParsed = allImagesLinks.map(image => {
+	// 			const parsed = image.replace(/\\/g, '');
 
-				const idRegexp = new RegExp('(?<=":)(.*?)(?=,")', 'g');
-				const id = parsed.match(idRegexp);
-				imagesIds.push(id);
+	// 			const idRegexp = new RegExp('(?<=":)(.*?)(?=,")', 'g');
+	// 			const id = parsed.match(idRegexp);
+	// 			imagesIds.push(id);
 
-				const urlRegexp = new RegExp(
-					'(?<=mediaURL":")(.*?)(?=",)',
-					'g'
-				);
-				const url = parsed.match(urlRegexp);
-				imagesLinks.push(url);
+	// 			const urlRegexp = new RegExp(
+	// 				'(?<=mediaURL":")(.*?)(?=",)',
+	// 				'g'
+	// 			);
+	// 			const url = parsed.match(urlRegexp);
+	// 			imagesLinks.push(url);
 
-				return null;
-			});
+	// 			return null;
+	// 		});
 
-			if (!isEmpty(allImagesLinksParsed)) {
-				let tempContent = parsedContent;
-				const imagesLinksUniq = uniq(imagesLinks);
-				const imagesIdsUniq = uniq(imagesIds);
-				let counter = imagesLinksUniq.length;
-				const checkCounter = imagesIdsUniq.length;
+	// 		if (!isEmpty(allImagesLinksParsed)) {
+	// 			let tempContent = parsedContent;
+	// 			const imagesLinksUniq = uniq(imagesLinks);
+	// 			const imagesIdsUniq = uniq(imagesIds);
+	// 			let counter = imagesLinksUniq.length;
+	// 			const checkCounter = imagesIdsUniq.length;
 
-				if (counter !== checkCounter) {
-					console.error(
-						__(
-							"Error processing images' links and ids - counts do not match",
-							'maxi-blocks'
-						)
-					);
-					replaceBlock(
-						clientId,
-						wp.blocks.rawHandler({
-							HTML: parsedContent,
-							mode: 'BLOCKS',
-						})
-					);
-					return;
-				}
+	// 			if (counter !== checkCounter) {
+	// 				console.error(
+	// 					__(
+	// 						"Error processing images' links and ids - counts do not match",
+	// 						'maxi-blocks'
+	// 					)
+	// 				);
+	// 				replaceBlock(
+	// 					clientId,
+	// 					wp.blocks.rawHandler({
+	// 						HTML: parsedContent,
+	// 						mode: 'BLOCKS',
+	// 					})
+	// 				);
+	// 				return;
+	// 			}
 
-				const imagesUniq = imagesIdsUniq.reduce(
-					(o, k, i) => ({ ...o, [k]: imagesLinksUniq[i] }),
-					{}
-				);
+	// 			const imagesUniq = imagesIdsUniq.reduce(
+	// 				(o, k, i) => ({ ...o, [k]: imagesLinksUniq[i] }),
+	// 				{}
+	// 			);
 
-				Object.entries(imagesUniq).map(image => {
-					const id = image[0];
-					const url = image[1];
+	// 			Object.entries(imagesUniq).map(image => {
+	// 				const id = image[0];
+	// 				const url = image[1];
 
-					imageUploader(url, usePlaceholderImage).then(data => {
-						tempContent = tempContent.replaceAll(url, data.url);
-						tempContent = tempContent.replaceAll(id, data.id);
-						counter -= 1;
-						if (counter === 0) {
-							replaceBlock(
-								clientId,
-								wp.blocks.rawHandler({
-									HTML: tempContent,
-									mode: 'BLOCKS',
-								})
-							);
-						}
-					});
-					return null;
-				});
-			} else {
-				// no images to process
-				replaceBlock(
-					clientId,
-					wp.blocks.rawHandler({
-						HTML: parsedContent,
-						mode: 'BLOCKS',
-					})
-				);
-			}
-		} else {
-			// not valid gutenberg code
-			// TODO: show a human-readable error here
-			console.error(__('The Code is not valid', 'maxi-blocks'));
-			onRequestClose();
-		}
-	};
+	// 				imageUploader(url, usePlaceholderImage).then(data => {
+	// 					tempContent = tempContent.replaceAll(url, data.url);
+	// 					tempContent = tempContent.replaceAll(id, data.id);
+	// 					counter -= 1;
+	// 					if (counter === 0) {
+	// 						replaceBlock(
+	// 							clientId,
+	// 							wp.blocks.rawHandler({
+	// 								HTML: tempContent,
+	// 								mode: 'BLOCKS',
+	// 							})
+	// 						);
+	// 					}
+	// 				});
+	// 				return null;
+	// 			});
+	// 		} else {
+	// 			// no images to process
+	// 			replaceBlock(
+	// 				clientId,
+	// 				wp.blocks.rawHandler({
+	// 					HTML: parsedContent,
+	// 					mode: 'BLOCKS',
+	// 				})
+	// 			);
+	// 		}
+	// 	} else {
+	// 		// not valid gutenberg code
+	// 		// TODO: show a human-readable error here
+	// 		console.error(__('The Code is not valid', 'maxi-blocks'));
+	// 		onRequestClose();
+	// 	}
+	// };
 
 	const getShapeType = type => {
 		switch (type) {
@@ -451,7 +452,15 @@ const LibraryContainer = props => {
 				taxonomies={hit.taxonomies?.category?.[0]}
 				serial={hit.post_number}
 				onRequestInsert={() =>
-					onRequestInsertPattern(hit.gutenberg_code, isChecked)
+					onRequestInsertPattern(
+						hit.gutenberg_code,
+						isChecked,
+						isValidTemplate,
+						onSelect,
+						onRequestClose,
+						replaceBlock,
+						clientId
+					)
 				}
 			/>
 		);
