+++ src/editor/library/editor.scss
@@ -117,6 +117,7 @@
 		margin-top: 0;
 		padding: 46px 0 0;
 		overflow-x: hidden;
+		overflow-y: clip;
 
 		&::before {
 			margin: 0;
