+++ src/editor/library/util.js
@@ -16,7 +16,7 @@ import {
 /**
  * External dependencies
  */
-import { isNil } from 'lodash';
+import { isNil, uniq, isEmpty } from 'lodash';
 
 export const rgbToHex = color => {
 	if (isNil(color)) return '';
@@ -233,4 +233,113 @@ export const svgCurrentColorStatus = (blockStyle, target = 'svg') => {
 					? iconInheritColor
 					: currentColor
 		  );
+};
+
+export const onRequestInsertPattern = (
+	parsedContent,
+	usePlaceholderImage,
+	isValidTemplate,
+	onSelect,
+	onRequestClose,
+	replaceBlock,
+	clientId
+) => {
+	const isValid = isValidTemplate(parsedContent);
+
+	if (isValid) {
+		const loadingMessage = `<h3>${__(
+			'LOADING…',
+			'maxi-blocks'
+		)}<span class="maxi-spinner"></span></h3>`;
+
+		onSelect({ content: loadingMessage });
+
+		onRequestClose();
+
+		const imagesLinks = [];
+		const imagesIds = [];
+
+		const allImagesRegexp = new RegExp('mediaID":(.*)",', 'g');
+
+		const allImagesLinks = parsedContent.match(allImagesRegexp);
+
+		const allImagesLinksParsed = allImagesLinks.map(image => {
+			const parsed = image.replace(/\\/g, '');
+
+			const idRegexp = new RegExp('(?<=":)(.*?)(?=,")', 'g');
+			const id = parsed.match(idRegexp);
+			imagesIds.push(id);
+
+			const urlRegexp = new RegExp('(?<=mediaURL":")(.*?)(?=",)', 'g');
+			const url = parsed.match(urlRegexp);
+			imagesLinks.push(url);
+
+			return null;
+		});
+
+		if (!isEmpty(allImagesLinksParsed)) {
+			let tempContent = parsedContent;
+			const imagesLinksUniq = uniq(imagesLinks);
+			const imagesIdsUniq = uniq(imagesIds);
+			let counter = imagesLinksUniq.length;
+			const checkCounter = imagesIdsUniq.length;
+
+			if (counter !== checkCounter) {
+				console.error(
+					__(
+						"Error processing images' links and ids - counts do not match",
+						'maxi-blocks'
+					)
+				);
+				replaceBlock(
+					clientId,
+					wp.blocks.rawHandler({
+						HTML: parsedContent,
+						mode: 'BLOCKS',
+					})
+				);
+				return;
+			}
+
+			const imagesUniq = imagesIdsUniq.reduce(
+				(o, k, i) => ({ ...o, [k]: imagesLinksUniq[i] }),
+				{}
+			);
+
+			Object.entries(imagesUniq).map(image => {
+				const id = image[0];
+				const url = image[1];
+
+				imageUploader(url, usePlaceholderImage).then(data => {
+					tempContent = tempContent.replaceAll(url, data.url);
+					tempContent = tempContent.replaceAll(id, data.id);
+					counter -= 1;
+					if (counter === 0) {
+						replaceBlock(
+							clientId,
+							wp.blocks.rawHandler({
+								HTML: tempContent,
+								mode: 'BLOCKS',
+							})
+						);
+					}
+				});
+				return null;
+			});
+		} else {
+			// no images to process
+			replaceBlock(
+				clientId,
+				wp.blocks.rawHandler({
+					HTML: parsedContent,
+					mode: 'BLOCKS',
+				})
+			);
+		}
+	} else {
+		// not valid gutenberg code
+		// TODO: show a human-readable error here
+		console.error(__('The Code is not valid', 'maxi-blocks'));
+		onRequestClose();
+	}
 };
